Move files, rename envs
