+++ .env.example
@@ -1,0 +1,2 @@
+DISCORD_TOKEN=
+TESTSERVER=
