--- src/Interfaces/config.ts
@@ -1,6 +1,0 @@
-declare namespace NodeJS {
-	export interface ProcessEnv {
-		TOKEN: string;
-		TESTSERVER: string;
-	}
-}
