+++ src/Interfaces/event.ts
@@ -1,4 +1,4 @@
-import Client from '../Client';
+import Client from '../lib/Client';
 import { ClientEvents } from 'discord.js';
 
 interface Run {
