+++ src/lib/types/node.d.ts
@@ -1,0 +1,8 @@
+declare namespace NodeJS {
+    export interface ProcessEnv {
+        TOKEN: string;
+        TESTSERVER: string;
+    }
+}
+
+export { }
